Remove "font-size: 100%" from CSS reset
As far as I know, the purpose of this was just to make all elements use
the same font size.  In practice, the only elements that would have a
different font size in some browsers are h1-h6, big, small, sup, sub,
and monospace elements (pre, code, kbd, samp, tt).  But the font size
for these is already specified elsewhere in the file (except for h1-h6,
for which it has been added with this commit).

Note that the html and body elements still have font-size specified.